fix refactoring errors, add coverage
