+++ tests/components/bluetooth/test_init.py
@@ -5,7 +5,6 @@ from unittest.mock import MagicMock, patch
 
 from bleak import BleakError
 from bleak.backends.scanner import AdvertisementData, BLEDevice
-from dbus_next import InvalidMessageError
 import pytest
 
 from homeassistant.components import bluetooth
@@ -20,12 +19,8 @@ from homeassistant.components.bluetooth import (
     scanner,
 )
 from homeassistant.components.bluetooth.const import (
-    CONF_ADAPTER,
-    SCANNER_WATCHDOG_INTERVAL,
-    SCANNER_WATCHDOG_TIMEOUT,
     SOURCE_LOCAL,
     UNAVAILABLE_TRACK_SECONDS,
-    UNIX_DEFAULT_BLUETOOTH_ADAPTER,
 )
 from homeassistant.config_entries import ConfigEntryState
 from homeassistant.const import EVENT_HOMEASSISTANT_STARTED, EVENT_HOMEASSISTANT_STOP
@@ -1495,208 +1490,3 @@ async def test_getting_the_scanner_returns_the_wrapped_instance(hass, enable_blu
     """Test getting the scanner returns the wrapped instance."""
     scanner = bluetooth.async_get_scanner(hass)
     assert isinstance(scanner, models.HaBleakScannerWrapper)
-
-
-async def test_config_entry_can_be_reloaded_when_stop_raises(
-    hass, caplog, enable_bluetooth
-):
-    """Test we can reload if stopping the scanner raises."""
-    entry = hass.config_entries.async_entries(bluetooth.DOMAIN)[0]
-    assert entry.state == ConfigEntryState.LOADED
-
-    with patch(
-        "homeassistant.components.bluetooth.scanner.OriginalBleakScanner.stop",
-        side_effect=BleakError,
-    ):
-        await hass.config_entries.async_reload(entry.entry_id)
-        await hass.async_block_till_done()
-
-    assert entry.state == ConfigEntryState.LOADED
-    assert "Error stopping scanner" in caplog.text
-
-
-async def test_changing_the_adapter_at_runtime(hass):
-    """Test we can change the adapter at runtime."""
-    entry = MockConfigEntry(
-        domain=bluetooth.DOMAIN,
-        data={},
-        options={CONF_ADAPTER: UNIX_DEFAULT_BLUETOOTH_ADAPTER},
-    )
-    entry.add_to_hass(hass)
-
-    with patch(
-        "homeassistant.components.bluetooth.scanner.OriginalBleakScanner.start"
-    ), patch("homeassistant.components.bluetooth.scanner.OriginalBleakScanner.stop"):
-        assert await hass.config_entries.async_setup(entry.entry_id)
-        await hass.async_block_till_done()
-
-        entry.options = {CONF_ADAPTER: "hci1"}
-
-        await hass.config_entries.async_reload(entry.entry_id)
-        await hass.async_block_till_done()
-
-        hass.bus.async_fire(EVENT_HOMEASSISTANT_STOP)
-        await hass.async_block_till_done()
-
-
-async def test_dbus_socket_missing_in_container(hass, caplog):
-    """Test we handle dbus being missing in the container."""
-
-    with patch(
-        "homeassistant.components.bluetooth.scanner.is_docker_env", return_value=True
-    ), patch(
-        "homeassistant.components.bluetooth.scanner.OriginalBleakScanner.start",
-        side_effect=FileNotFoundError,
-    ):
-        assert await async_setup_component(
-            hass, bluetooth.DOMAIN, {bluetooth.DOMAIN: {}}
-        )
-        await hass.async_block_till_done()
-        hass.bus.async_fire(EVENT_HOMEASSISTANT_STARTED)
-        await hass.async_block_till_done()
-
-    hass.bus.async_fire(EVENT_HOMEASSISTANT_STOP)
-    await hass.async_block_till_done()
-    assert "/run/dbus" in caplog.text
-    assert "docker" in caplog.text
-
-
-async def test_dbus_socket_missing(hass, caplog):
-    """Test we handle dbus being missing."""
-
-    with patch(
-        "homeassistant.components.bluetooth.scanner.is_docker_env", return_value=False
-    ), patch(
-        "homeassistant.components.bluetooth.scanner.OriginalBleakScanner.start",
-        side_effect=FileNotFoundError,
-    ):
-        assert await async_setup_component(
-            hass, bluetooth.DOMAIN, {bluetooth.DOMAIN: {}}
-        )
-        await hass.async_block_till_done()
-        hass.bus.async_fire(EVENT_HOMEASSISTANT_STARTED)
-        await hass.async_block_till_done()
-
-    hass.bus.async_fire(EVENT_HOMEASSISTANT_STOP)
-    await hass.async_block_till_done()
-    assert "DBus" in caplog.text
-    assert "docker" not in caplog.text
-
-
-async def test_dbus_broken_pipe_in_container(hass, caplog):
-    """Test we handle dbus broken pipe in the container."""
-
-    with patch(
-        "homeassistant.components.bluetooth.scanner.is_docker_env", return_value=True
-    ), patch(
-        "homeassistant.components.bluetooth.scanner.OriginalBleakScanner.start",
-        side_effect=BrokenPipeError,
-    ):
-        assert await async_setup_component(
-            hass, bluetooth.DOMAIN, {bluetooth.DOMAIN: {}}
-        )
-        await hass.async_block_till_done()
-        hass.bus.async_fire(EVENT_HOMEASSISTANT_STARTED)
-        await hass.async_block_till_done()
-
-    hass.bus.async_fire(EVENT_HOMEASSISTANT_STOP)
-    await hass.async_block_till_done()
-    assert "dbus" in caplog.text
-    assert "restarting" in caplog.text
-    assert "container" in caplog.text
-
-
-async def test_dbus_broken_pipe(hass, caplog):
-    """Test we handle dbus broken pipe."""
-
-    with patch(
-        "homeassistant.components.bluetooth.scanner.is_docker_env", return_value=False
-    ), patch(
-        "homeassistant.components.bluetooth.scanner.OriginalBleakScanner.start",
-        side_effect=BrokenPipeError,
-    ):
-        assert await async_setup_component(
-            hass, bluetooth.DOMAIN, {bluetooth.DOMAIN: {}}
-        )
-        await hass.async_block_till_done()
-        hass.bus.async_fire(EVENT_HOMEASSISTANT_STARTED)
-        await hass.async_block_till_done()
-
-    hass.bus.async_fire(EVENT_HOMEASSISTANT_STOP)
-    await hass.async_block_till_done()
-    assert "DBus" in caplog.text
-    assert "restarting" in caplog.text
-    assert "container" not in caplog.text
-
-
-async def test_invalid_dbus_message(hass, caplog):
-    """Test we handle invalid dbus message."""
-
-    with patch(
-        "homeassistant.components.bluetooth.scanner.OriginalBleakScanner.start",
-        side_effect=InvalidMessageError,
-    ):
-        assert await async_setup_component(
-            hass, bluetooth.DOMAIN, {bluetooth.DOMAIN: {}}
-        )
-        await hass.async_block_till_done()
-        hass.bus.async_fire(EVENT_HOMEASSISTANT_STARTED)
-        await hass.async_block_till_done()
-
-    hass.bus.async_fire(EVENT_HOMEASSISTANT_STOP)
-    await hass.async_block_till_done()
-    assert "dbus" in caplog.text
-
-
-async def test_recovery_from_dbus_restart(
-    hass, mock_bleak_scanner_start, enable_bluetooth
-):
-    """Test we can recover when DBus gets restarted out from under us."""
-    assert await async_setup_component(hass, bluetooth.DOMAIN, {bluetooth.DOMAIN: {}})
-    await hass.async_block_till_done()
-    assert len(mock_bleak_scanner_start.mock_calls) == 1
-
-    start_time_monotonic = 1000
-    scanner = _get_manager()
-    mock_discovered = [MagicMock()]
-    mock_discovered_devices(mock_discovered)
-
-    # Ensure we don't restart the scanner if we don't need to
-    with patch(
-        "homeassistant.components.bluetooth.scanner.MONOTONIC_TIME",
-        return_value=start_time_monotonic + 10,
-    ):
-        async_fire_time_changed(hass, dt_util.utcnow() + SCANNER_WATCHDOG_INTERVAL)
-        await hass.async_block_till_done()
-
-    assert len(mock_bleak_scanner_start.mock_calls) == 1
-
-    # Fire a callback to reset the timer
-    with patch(
-        "homeassistant.components.bluetooth.scanner.MONOTONIC_TIME",
-        return_value=start_time_monotonic,
-    ):
-        scanner._callback(
-            BLEDevice("44:44:33:11:23:42", "any_name"),
-            AdvertisementData(local_name="any_name"),
-        )
-
-    # Ensure we don't restart the scanner if we don't need to
-    with patch(
-        "homeassistant.components.bluetooth.scanner.MONOTONIC_TIME",
-        return_value=start_time_monotonic + 20,
-    ):
-        async_fire_time_changed(hass, dt_util.utcnow() + SCANNER_WATCHDOG_INTERVAL)
-        await hass.async_block_till_done()
-
-    assert len(mock_bleak_scanner_start.mock_calls) == 1
-
-    # We hit the timer, so we restart the scanner
-    with patch(
-        "homeassistant.components.bluetooth.scanner.MONOTONIC_TIME",
-        return_value=start_time_monotonic + SCANNER_WATCHDOG_TIMEOUT,
-    ):
-        async_fire_time_changed(hass, dt_util.utcnow() + SCANNER_WATCHDOG_INTERVAL)
-        await hass.async_block_till_done()
-
-    assert len(mock_bleak_scanner_start.mock_calls) == 2
